renaming and change folder for sight model
