+++ lib/data/interactor/sight_interactor.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:places/data/model/coordinate.dart';
 import 'package:places/data/model/enums/sight_type.dart';
 import 'package:places/data/model/sights/sights_filter_request_dto.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/sights/sight.dart';
 import 'package:places/data/repository/sight_repository.dart';
 import 'package:places/data/services/location_service.dart';
