+++ lib/data/model/sights/sight_dto.dart
@@ -1,9 +1,9 @@
 import 'package:json_annotation/json_annotation.dart';
 
 import '../enums/sight_type.dart';
-import '../sight.dart';
+import 'sight.dart';
 
-part '../sights/sight_dto.g.dart';
+part 'sight_dto.g.dart';
 
 @JsonSerializable()
 class SightDto extends Sight {
