+++ lib/ui/screen/visiting_screen/draggable_sight_cards_list_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/sights/sight.dart';
 import 'package:places/ui/widgets/sight_cards/sight_card_dismissible.dart';
 import 'package:places/ui/widgets/sight_cards/sight_card_imitation.dart';
 import 'package:places/ui/widgets/sight_cards/visited_sight_card.dart';
