+++ lib/ui/widgets/sight_cards/sight_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/sights/sight.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/widgets/sight_cards/sight_card_base.dart';
 import 'package:places/ui/widgets/sight_cards/sight_card_icon_button.dart';
