+++ lib/ui/widgets/sight_cards/want_to_visit_sight_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:places/constants/app_strings.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/sights/sight.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/widgets/pickers.dart';
 import 'package:places/ui/widgets/sight_cards/sight_card_dismissible.dart';
